requirements.txt
disnake
pillow
requests